% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbreg.R
\name{choose_strategy}
\alias{choose_strategy}
\title{Choose regression strategy based on inputs and auto logic}
\usage{
choose_strategy(inputs)
}
\description{
Choose regression strategy based on inputs and auto logic
}
\keyword{internal}
