% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbreg.R
\name{compute_vcov}
\alias{compute_vcov}
\title{Compute variance-covariance matrix}
\usage{
compute_vcov(
  vcov_type = "iid",
  strategy = "compress",
  XtX_inv,
  rss,
  df_res,
  nobs_orig,
  X = NULL,
  rss_g = NULL
)
}
\description{
Compute variance-covariance matrix
}
\keyword{internal}
