% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckreg.R
\name{execute_compress_strategy}
\alias{execute_compress_strategy}
\title{Execute compress strategy (groupby compression)}
\usage{
execute_compress_strategy(inputs)
}
\description{
Execute compress strategy (groupby compression)
}
\keyword{internal}
