% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckreg.R
\name{execute_group_strategy}
\alias{execute_group_strategy}
\title{Execute group strategy (groupby compression)}
\usage{
execute_group_strategy(inputs)
}
\description{
Execute group strategy (groupby compression)
}
\keyword{internal}
