% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckreg.R
\name{execute_moments_fe_strategy}
\alias{execute_moments_fe_strategy}
\title{Execute moments_fe strategy (1-2 fixed effects)}
\usage{
execute_moments_fe_strategy(inputs)
}
\description{
Execute moments_fe strategy (1-2 fixed effects)
}
\keyword{internal}
