% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbreg.R
\name{finalize_dbreg_result}
\alias{finalize_dbreg_result}
\title{Finalize dbreg result object}
\usage{
finalize_dbreg_result(result, inputs, chosen_strategy)
}
\description{
Finalize dbreg result object
}
\keyword{internal}
