% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckreg.R
\name{finalize_duckreg_result}
\alias{finalize_duckreg_result}
\title{Finalize duckreg result object}
\usage{
finalize_duckreg_result(result, inputs, chosen_strategy)
}
\description{
Finalize duckreg result object
}
\keyword{internal}
