% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbreg.R
\name{gen_coeftable}
\alias{gen_coeftable}
\title{Generate coefficient table from estimates and vcov matrix}
\usage{
gen_coeftable(betahat, vcov_mat, df_residual)
}
\description{
Generate coefficient table from estimates and vcov matrix
}
\keyword{internal}
