% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbreg.R
\name{gen_xvar_pairs}
\alias{gen_xvar_pairs}
\title{Generate unique pairs of variables (preserves original nested loop order)}
\usage{
gen_xvar_pairs(xvars)
}
\description{
Generate unique pairs of variables (preserves original nested loop order)
}
\keyword{internal}
