% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.duckreg}
\alias{print.duckreg}
\title{Print method for duckreg objects}
\usage{
\method{print}{duckreg}(x, fes = FALSE, ...)
}
\arguments{
\item{x}{`duckreg` object.}

\item{fes}{Should the fixed effects be displayed? Default is `FALSE`.}

\item{...}{Other arguments passed to \code{\link[base]{print}}. Not currently used.}
}
\description{
Print method for duckreg objects
}
