% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckreg.R
\name{process_duckreg_inputs}
\alias{process_duckreg_inputs}
\title{Process and validate duckreg inputs}
\usage{
process_duckreg_inputs(
  fml,
  conn,
  table,
  data,
  path,
  vcov,
  strategy,
  query_only,
  data_only,
  compression_threshold,
  verbose,
  ridge_rel
)
}
\description{
Process and validate duckreg inputs
}
\keyword{internal}
