% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbreg.R
\name{solve_with_ridge_fallback}
\alias{solve_with_ridge_fallback}
\title{Solve linear system with ridge fallback}
\usage{
solve_with_ridge_fallback(XtX, Xty, ridge_rel)
}
\description{
Solve linear system with ridge fallback
}
\keyword{internal}
