% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckreg.R
\name{duckreg}
\alias{duckreg}
\title{Run a compressed regression with a DuckDB backend.}
\usage{
duckreg(fml, conn = NULL, table = NULL, data = NULL, path = NULL, vcov = "hc1")
}
\arguments{
\item{fml}{A \code{\link[stats]{formula}} representing the relation to be
estimated. Fixed-effects should be included after a pipe, e.g
\code{fml = y ~ x1 + x2 | fe1 + f2}. Currently, only simple additive terms
are supported (i.e., no interaction terms, transformations or literals).}

\item{conn}{Connection to a DuckDB database, e.g. created with
\code{\link[DBI]{dbConnect}}. Can be either persistent (disk-backed) or
ephemeral (in-memory). If no connection is provided, then an ephemeral
connection will be created automatically and closed before the function
exits. Note that a persistent (disk-backed) database connection is
required for larger-than-RAM datasets in order to take advantage of DuckDB's
streaming functionality.}

\item{table, data, path}{Mututally exclusive arguments for specifying the data
table (object) to be queried. In order of precedence:
\itemize{
\item \code{table}: Character string giving the name of the data table in an
existing (open) DuckDB connection.
\item \code{data}: R dataframe that can be copied over to \code{conn} as a temporary
table for querying via the DuckDB query engine. Ignored if \code{table} is
provided.
\item \code{path}: Character string giving a path to the data file(s) on disk, which
will be read into \code{conn}. Internally, this string is passed to the \code{FROM}
query statement, so could (should) include file globbing for
Hive-partitioned datasets, e.g. \code{"mydata/**/.*parquet"}. For more precision,
however, it is recommended to pass the desired DuckDB reader function as
part of this string, e.g. \code{"read_parquet('mydata/**/*.parquet')"};
note the use of single quotes.
Ignored if either \code{table} or \code{data} is provided.
}}

\item{vcov}{Character string denoting the desired type of variance-
covariance correction / standard errors. At present, only "hc1"
(heteroskedasticity-consistent) are supported, which is also thus
the default.}
}
\value{
A two-column matrix containing the regression results:
\itemize{
\item estimates: Point estimates
\item std.error: Standard errors
}
}
\description{
Run a compressed regression with a DuckDB backend.
}
\examples{
1+1
}
