% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckreg.R
\name{execute_moments_strategy}
\alias{execute_moments_strategy}
\title{Execute moments strategy (no fixed effects)}
\usage{
execute_moments_strategy(inputs)
}
\description{
Execute moments strategy (no fixed effects)
}
\keyword{internal}
